int-as-pointer parameters: 12 more slots, and four offsets read through their fields
Twelve more of the class, plus the raw byte arithmetic three of them carried.

  debugPrint (3)  debugPrintDraw, debugPrintDrawRecord and debugPrintDrawGlyph
      all take void* context.  The evidence is the one consumer,
      setTextColor(void* context, ...), whose other callers pass &gdl -- a
      display list, not an object -- so the slot takes the callee's own
      parameter name rather than the misleading `color` it had.  gameloop_main's
      debugPrintDraw(0) needs no change.
  cacheAllocAndCopy  u8* srcAddress; the body's -= alignOffset / += 0x1000 walk
      is now byte arithmetic on a byte pointer instead of on a u32, and the four
      track_dolphin call sites cast to u8* instead of u32.
  piRomLoadSection  void* destBuf; shader.c's only call drops its (int) and
      passes MapRomListPage.objects, which is already an ObjPlacement*.
  Kaldachom / DLL 0xCE (2)  kaldachom_updateCombat and dll_CE_updateTargeting
      take GroundBaddieState*, and their
      processMessages(..., (void*)(objectStateAddress + 0x35c), ...) becomes
      &objectStateAddress->routeNav -- offsetof(GroundBaddieState, routeNav) is
      STATIC_ASSERTed at 0x35C, and both files ALREADY spell the neighbouring
      updateHitReaction/startHitReaction argument that way, so this is the file's
      own idiom, not a new guess.  Both callers passed the same pointer twice,
      once as (int) and once typed; now they pass it twice typed.
  SHthorntail (2)  updateTailSwing and updateLevelControlMode1 take GameObject*.
  DIM boss  the three DLL-local Dim2*Fn interface typedefs and
      DIMbossAnim_updateBossHitReaction take GameObject*, closing the last
      int obj in that unit's handler tables.
  DBstealerworm (2)  processEffectFlags and acquireTarget take
      GroundBaddieState*.
  explodable_buildFragments  ExplodableState*, and the alias that existed only
      to carry the cast is gone.

Measured and NOT landed, reverted whole:
  dll500_update  .text differs at identical size -- the parameter is used both
      as a pointer and as an int, so once it is a pointer MWCC re-colours.
DECLINED on the evidence, not on bytes:
  drshackle_renderAtPathPoint's `a`  is REASSIGNED mid-body (a = (int)p) and
      then walked with a += 4: two objects in one slot.
  modelLoadMtxsToGx's `obj`  is likewise reassigned, to a joint COUNT.
  textureIdxToPtr's `idx`  is a tagged handle by design -- the high bit selects
      between an already-resolved pointer and a one-based index.

GATE, twice, build-idle, against a pristine control built at the same parent:
1005 objects compared / 0 differ / 0 missing / 0 new; 2 carry only @NNN
literal-pool renumbering at preserved addresses.  Tree unchanged at fuzzy
99.81484 / matched_code 2513248 / matched_data 1195969 / complete_units 914 of
1043.  banned_shapes hits=103 identical to the control.  ninja all_source
EXIT=0; forced link main.dol: OK, sha1 e750e8e894707a52446118a4b84f1b58b677b269
md5 7b955850ea4bd7ceda0109493203ff5b.

+++ src/main/shader.c
@@ -3086,7 +3086,7 @@ int mapGetRomListAndOffsets(int p1, int flag)
     ((MapRomListPage*)gCurRomListPage)->visLayerRects = (u32*)((int)gCurRomListPage + *(int*)((gMapsTab + 0x14) + (words << 2)) - offset0);
     ((MapRomListPage*)gCurRomListPage)->objects = (ObjPlacement*)((int)gCurRomListPage + *(int*)((gMapsTab + 0x18) + (words << 2)) - offset0);
 
-    piRomLoadSection(*(int*)((gMapsTab + 0x18) + (words << 2)), p1, (int)((MapRomListPage*)gCurRomListPage)->objects);
+    piRomLoadSection(*(int*)((gMapsTab + 0x18) + (words << 2)), p1, ((MapRomListPage*)gCurRomListPage)->objects);
     ((MapRomListPage*)gCurRomListPage)->loadedObjectBits = (u8*)((*(int*)((gMapsTab + 0x1c) + (words << 2)) + v2) + (int)gCurRomListPage - offset0);
 
     for (i = 0; i < (v0 + 7 >> 3) + 1; i++)
